requirements.txt
asgiref==3.7.2
certifi==2020.6.20
chardet==3.0.4
decouple==0.0.7
defusedxml==0.6.0
Django==4.2.6
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.0.4
idna==2.10
isort==5.12.0
oauthlib==3.1.0
packaging==23.2
Pillow==8.1.1
PyJWT==1.7.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==6.6.0
